requirements.aws.txt
aws-cdk.core
aws-cdk.aws_lambda